+++ test/inputs/collection_radio_buttons_input_test.rb
@@ -186,12 +186,6 @@ class CollectionRadioButtonsInputTest < ActionView::TestCase
     assert_select 'input[type=radio][required]'
   end
 
-  test 'collection input with radio type generates aria-required html attribute' do
-    with_input_for @user, :name, :radio_buttons, collection: %w[Jose Carlos]
-    assert_select 'input[type=radio].required'
-    assert_select 'input[type=radio][aria-required=true]'
-  end
-
   test 'input radio does not wrap the collection by default' do
     with_input_for @user, :active, :radio_buttons