+++ test/inputs/collection_select_input_test.rb
@@ -283,14 +283,12 @@ class CollectionSelectInputTest < ActionView::TestCase
   test "collection input generated aria-label should contain 'true'" do
     with_input_for @user, :age, :select, collection: 18..30, prompt: "Please select foo"
     assert_select 'select.required'
-    assert_select 'select[aria-required=true]'
   end
 
   test 'collection input with select type does not generate required html attribute without blank option' do
     with_input_for @user, :name, :select, include_blank: false, collection: %w[Jose Carlos]
     assert_select 'select.required'
     assert_no_select 'select[required]'
-    assert_no_select 'select[aria-required=true]'
   end
 
   test 'collection input with select type with multiple attribute generates required html attribute without blank option' do
@@ -303,30 +301,6 @@ class CollectionSelectInputTest < ActionView::TestCase
     with_input_for @user, :name, :select, include_blank: true, input_html: { multiple: true }, collection: %w[Jose Carlos]
     assert_select 'select.required'
     assert_select 'select[required]'
-  end
-
-  test 'with a blank option, a collection input of type select has an aria-required html attribute' do
-    with_input_for @user, :name, :select, include_blank: true, collection: %w[Jose Carlos]
-    assert_select 'select.required'
-    assert_select 'select[aria-required=true]'
-  end
-
-  test 'without a blank option, a collection input of type select does not have an aria-required html attribute' do
-    with_input_for @user, :name, :select, include_blank: false, collection: %w[Jose Carlos]
-    assert_select 'select.required'
-    assert_no_select 'select[aria-required]'
-  end
-
-  test 'without a blank option and with a multiple option, a collection input of type select has an aria-required html attribute' do
-    with_input_for @user, :name, :select, include_blank: false, input_html: { multiple: true }, collection: %w[Jose Carlos]
-    assert_select 'select.required'
-    assert_select 'select[aria-required=true]'
-  end
-
-  test 'with a blank option and a multiple option, a collection input of type select has an aria-required html attribute' do
-    with_input_for @user, :name, :select, include_blank: true, input_html: { multiple: true }, collection: %w[Jose Carlos]
-    assert_select 'select.required'
-    assert_select 'select[aria-required]'
   end
 
   test 'input allows disabled options with a lambda for collection select' do