+++ test/inputs/country_input_test.rb
@@ -27,10 +27,4 @@ class CountryInputTest < ActionView::TestCase
     assert_select 'select.required'
     assert_select 'select[required]'
   end
-
-  test 'input does generate select element with aria-required html attribute' do
-    with_input_for @user, :country, :country
-    assert_select 'select.required'
-    assert_select 'select[aria-required]'
-  end
 end